+++ src/hooks/filteredTodos/useFilteredTodosState.js
@@ -1,7 +1,7 @@
-import useTodosState from "./useTodosState";
+// 필터링된 할 일 목록과 관련된 상태를 관리하는 훅
+import useTodosState from "../todos/useTodosState";
 import useFilteredTodos from "./useFilteredTodos";
 
-// 필터링된 할 일 목록과 관련된 상태를 관리하는 훅
 const useFilteredTodosState = () => {
   const { todos, setTodos, filter, setFilter } = useTodosState();
   const filteredTodos = useFilteredTodos(todos, filter);
