+++ src/hooks/filters/useFilterState.js
@@ -1,7 +1,7 @@
-import { useState } from "react";
-import FILTERS from "../utils/filters";
-
 // 필터 상태를 관리하는 훅
+import { useState } from "react";
+import FILTERS from "../../utils/filters";
+
 const useFilterState = () => {
   const [filter, setFilter] = useState(FILTERS.ALL);
