+++ src/hooks/storage/useSaveTodos.js
@@ -1,13 +1,10 @@
+// 로컬 스토리지에 할 일 목록을 저장하는 훅
 import { useEffect } from "react";
 
-const useSaveTodos = (allTodos, filter) => {
+const useSaveTodos = (allTodos) => {
   useEffect(() => {
     localStorage.setItem("todos", JSON.stringify(allTodos));
   }, [allTodos]);
-
-  useEffect(() => {
-    localStorage.setItem("filter", filter);
-  }, [filter]);
 };
 
 export default useSaveTodos;
