+++ src/hooks/todos/useTodoList.js
@@ -1,6 +1,6 @@
+// 할 일 목록 상태를 관리하는 훅
 import { useState } from "react";
 
-// 할 일 목록 상태를 관리하는 훅
 const useTodoList = () => {
   const [todos, setTodos] = useState([]);
