+++ src/hooks/useTodoHandlers.js
@@ -1,3 +1,4 @@
+// 할 일 관련 핸들러를 관리하는 훅
 import {
   handleAddTodo,
   handleToggleComplete,
@@ -5,14 +6,11 @@ import {
   handleSetFilter,
 } from "../handlers";
 
-const useTodoHandlers = (todos, setTodos, setFilter) => {
-  return {
-    handleAddTodo: (todo) => handleAddTodo(todos, setTodos, todo),
-    handleToggleComplete: (index) =>
-      handleToggleComplete(todos, setTodos, index),
-    handleDeleteTodo: (index) => handleDeleteTodo(todos, setTodos, index),
-    handleSetFilter: (filter) => handleSetFilter(setFilter, filter),
-  };
-};
+const useTodoHandlers = (todos, setTodos, setFilter) => ({
+  handleAddTodo: (todo) => handleAddTodo(todos, setTodos, todo),
+  handleToggleComplete: (index) => handleToggleComplete(todos, setTodos, index),
+  handleDeleteTodo: (index) => handleDeleteTodo(todos, setTodos, index),
+  handleSetFilter: (filter) => handleSetFilter(setFilter, filter),
+});
 
 export default useTodoHandlers;
